🚚 [routers.local_cdn] => move `emit_local_cdn_status`
